Moved feeder into a new file; made JSON parsing more robust
